move register middleware function into signupController
