+++ server/controller/loginController.js
@@ -1,0 +1,63 @@
+const express = require('express');
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+const passport = require('passport');
+
+const User = require("../models/userModel");
+const loginController = {};
+// Load input validation
+const validateLoginInput = require("../validation/login");
+
+// @route POST api/users/login
+// @desc Login user and return JWT token
+// @access Public
+//validate user inputs and send appropriate errors when necessary
+loginController.logInUser = (req, res, next) => {
+    // Form validation
+    const { errors, isValid } = validateLoginInput(req.body);
+    // Check validation
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
+    const email = req.body.email;
+    const password = req.body.password;
+    // Find user by email
+    User.findOne({ email }).then(user => {
+        // Check if user exists
+        if (!user) {
+            return res.status(404).json({ emailnotfound: "Email not found" });
+        }
+        // Check password
+        bcrypt.compare(password, user.password).then(isMatch => {
+            if (isMatch) {
+                // User matched
+                // Create JWT Payload
+                const payload = {
+                    id: user.id,
+                    name: user.name
+                };
+                // Sign token
+                jwt.sign(
+                    payload,
+                    keys.secretOrKey,
+                    {
+                        expiresIn: 31556926 // 1 year in seconds
+                    },
+                    (err, token) => {
+                        res.json({
+                            success: true,
+                            token: "Bearer " + token
+                        });
+                    }
+                );
+            } else {
+                return res
+                    .status(400)
+                    .json({ passwordincorrect: "Password incorrect" });
+            }
+        });
+    });
+};
+
+
+module.exports = loginController;
