+++ server/controller/signupController.js
@@ -1,0 +1,45 @@
+const express = require('express');
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+const passport = require('passport');
+const User = require("../models/userModel");
+const signupController = {};
+// Load input validation
+const validateRegisterInput = require("../validation/register");
+
+// @route POST api/users/register
+// @desc Register user
+// @access Public
+signupController.registerUser = (req, res, next) => {
+    // Form validation
+    console.log('req.body: ', req.body)
+    const { errors, isValid } = validateRegisterInput(req.body);
+    // Check validation
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
+    User.findOne({ email: req.body.email }).then(user => {
+        if (user) {
+            return res.status(400).json({ email: "Email already exists" });
+        } else {
+            const newUser = new User({
+                name: req.body.name,
+                email: req.body.email,
+                password: req.body.password
+            });
+            // Hash password before saving in database
+            bcrypt.genSalt(10, (err, salt) => {
+                bcrypt.hash(newUser.password, salt, (err, hash) => {
+                    if (err) throw err;
+                    newUser.password = hash;
+                    newUser
+                        .save()
+                        .then(user => res.json(user))
+                        .catch(err => console.log(err));
+                });
+            });
+        }
+    });
+}
+
+module.exports = signupController;
