+++ server/routes/users.js
@@ -15,87 +15,87 @@ const User = require("../models/userModel");
 // @route POST api/users/register
 // @desc Register user
 // @access Public
-usersRouter.post("/register", (req, res) => {
-    // Form validation
-    console.log('req.body: ', req.body)
-    const { errors, isValid } = validateRegisterInput(req.body);
-    // Check validation
-    if (!isValid) {
-        return res.status(400).json(errors);
-    }
-    User.findOne({ email: req.body.email }).then(user => {
-        if (user) {
-            return res.status(400).json({ email: "Email already exists" });
-        } else {
-            const newUser = new User({
-                name: req.body.name,
-                email: req.body.email,
-                password: req.body.password
-            });
-            // Hash password before saving in database
-            bcrypt.genSalt(10, (err, salt) => {
-                bcrypt.hash(newUser.password, salt, (err, hash) => {
-                    if (err) throw err;
-                    newUser.password = hash;
-                    newUser
-                        .save()
-                        .then(user => res.json(user))
-                        .catch(err => console.log(err));
-                });
-            });
-        }
-    });
-});
+// usersRouter.post("/register", (req, res) => {
+//     // Form validation
+//     console.log('req.body: ', req.body)
+//     const { errors, isValid } = validateRegisterInput(req.body);
+//     // Check validation
+//     if (!isValid) {
+//         return res.status(400).json(errors);
+//     }
+//     User.findOne({ email: req.body.email }).then(user => {
+//         if (user) {
+//             return res.status(400).json({ email: "Email already exists" });
+//         } else {
+//             const newUser = new User({
+//                 name: req.body.name,
+//                 email: req.body.email,
+//                 password: req.body.password
+//             });
+//             // Hash password before saving in database
+//             bcrypt.genSalt(10, (err, salt) => {
+//                 bcrypt.hash(newUser.password, salt, (err, hash) => {
+//                     if (err) throw err;
+//                     newUser.password = hash;
+//                     newUser
+//                         .save()
+//                         .then(user => res.json(user))
+//                         .catch(err => console.log(err));
+//                 });
+//             });
+//         }
+//     });
+// });
 
 // @route POST api/users/login
 // @desc Login user and return JWT token
 // @access Public
-usersRouter.post("/login", (req, res) => {
-    // Form validation
-    const { errors, isValid } = validateLoginInput(req.body);
-    // Check validation
-    if (!isValid) {
-        return res.status(400).json(errors);
-    }
-    const email = req.body.email;
-    const password = req.body.password;
-    // Find user by email
-    User.findOne({ email }).then(user => {
-        // Check if user exists
-        if (!user) {
-            return res.status(404).json({ emailnotfound: "Email not found" });
-        }
-        // Check password
-        bcrypt.compare(password, user.password).then(isMatch => {
-            if (isMatch) {
-                // User matched
-                // Create JWT Payload
-                const payload = {
-                    id: user.id,
-                    name: user.name
-                };
-                // Sign token
-                jwt.sign(
-                    payload,
-                    keys.secretOrKey,
-                    {
-                        expiresIn: 31556926 // 1 year in seconds
-                    },
-                    (err, token) => {
-                        res.json({
-                            success: true,
-                            token: "Bearer " + token
-                        });
-                    }
-                );
-            } else {
-                return res
-                    .status(400)
-                    .json({ passwordincorrect: "Password incorrect" });
-            }
-        });
-    });
-});
+// usersRouter.post("/login", (req, res) => {
+//     // Form validation
+//     const { errors, isValid } = validateLoginInput(req.body);
+//     // Check validation
+//     if (!isValid) {
+//         return res.status(400).json(errors);
+//     }
+//     const email = req.body.email;
+//     const password = req.body.password;
+//     // Find user by email
+//     User.findOne({ email }).then(user => {
+//         // Check if user exists
+//         if (!user) {
+//             return res.status(404).json({ emailnotfound: "Email not found" });
+//         }
+//         // Check password
+//         bcrypt.compare(password, user.password).then(isMatch => {
+//             if (isMatch) {
+//                 // User matched
+//                 // Create JWT Payload
+//                 const payload = {
+//                     id: user.id,
+//                     name: user.name
+//                 };
+//                 // Sign token
+//                 jwt.sign(
+//                     payload,
+//                     keys.secretOrKey,
+//                     {
+//                         expiresIn: 31556926 // 1 year in seconds
+//                     },
+//                     (err, token) => {
+//                         res.json({
+//                             success: true,
+//                             token: "Bearer " + token
+//                         });
+//                     }
+//                 );
+//             } else {
+//                 return res
+//                     .status(400)
+//                     .json({ passwordincorrect: "Password incorrect" });
+//             }
+//         });
+//     });
+// });
 
 //route to hit for google oauth
 usersRouter.get('/google',
